requirements.txt
Flask
openpyxl
google-genai
dotenv